refactored postDeletePRoduct + postEditProduct shop controller, delete image files when removed from db

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
 
@@ -19,21 +21,21 @@ exports.postAddProduct = (req, res, next) => {
   const description = req.body.description;
   if (!image) {
     return res.status(422).
-    render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
-      editing: false,
-      hasError: true,
-      product: {
-        title: title,
-        price: price,
-        description: description
-      },
-      errorMessage: 'Attached file is not an image!',
-      validationErrors: []
-    });
+      render('admin/edit-product', {
+        pageTitle: 'Add Product',
+        path: '/admin/add-product',
+        editing: false,
+        hasError: true,
+        product: {
+          title: title,
+          price: price,
+          description: description
+        },
+        errorMessage: 'Attached file is not an image!',
+        validationErrors: []
+      });
   }
-  
+
   const imageUrl = image.path;
 
   const errors = validationResult(req);
@@ -49,7 +51,7 @@ exports.postAddProduct = (req, res, next) => {
           price: price,
           description: description
         },
-        errorMessage:'Attached file is not an image!',
+        errorMessage: 'Attached file is not an image!',
         validationErrors: []
       });
   }
@@ -137,6 +139,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       product.save()
@@ -171,7 +174,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found!'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(result => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
